mcp_server/requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.0
asyncio==3.4.3